jaeger_ui_auth/requirements.txt
python-dotenv==1.0.1
gunicorn==23.0.0
Flask==3.1.0
PyJWT==2.10.1
requests==2.32.3

